Reorganised some files
