+++ CodeEdit/Features/SplitView/SplitViewControllerView.swift
@@ -6,47 +6,8 @@
 //
 
 import SwiftUI
-import Combine
 
-class SplitViewItem: ObservableObject {
-
-    var id: AnyHashable
-    var item: NSSplitViewItem
-
-    var collapsed: Binding<Bool>
-
-    var cancellables: [AnyCancellable] = []
-
-    var observers: [NSKeyValueObservation] = []
-
-    init(child: _VariadicView.Children.Element) {
-        self.id = child.id
-        self.item = NSSplitViewItem(viewController: NSHostingController(rootView: child))
-        self.collapsed = child[SplitViewItemCollapsedViewTraitKey.self]
-        self.item.canCollapse = child[SplitViewItemCanCollapseViewTraitKey.self]
-        self.item.isCollapsed = self.collapsed.wrappedValue
-        self.observers = createObservers()
-    }
-
-    func createObservers() -> [NSKeyValueObservation] {
-        [
-            item.observe(\.isCollapsed) { item, _ in
-                self.collapsed.wrappedValue = item.isCollapsed
-            }
-        ]
-    }
-
-    func update(child: _VariadicView.Children.Element) {
-        self.item.canCollapse = child[SplitViewItemCanCollapseViewTraitKey.self]
-        DispatchQueue.main.async {
-            self.observers = []
-            self.item.animator().isCollapsed = child[SplitViewItemCollapsedViewTraitKey.self].wrappedValue
-            self.observers = self.createObservers()
-        }
-    }
-}
-
-struct EditorSplitView: NSViewControllerRepresentable {
+struct SplitViewControllerView: NSViewControllerRepresentable {
 
     var children: _VariadicView.Children
     var viewController: SplitViewController
@@ -125,8 +86,4 @@ final class SplitViewController: NSSplitViewController {
     func collapse(for id: AnyHashable, enabled: Bool) {
         items.first { $0.id == id }?.item.animator().isCollapsed = enabled
     }
-
-//    override func splitViewDidResizeSubviews(_ notification: Notification) {
-//        print(notification)
-//    }
 }
