+++ CodeEdit/Features/SplitView/SplitViewItem.swift
@@ -1,0 +1,47 @@
+//
+//  SplitViewItem.swift
+//  CodeEdit
+//
+//  Created by Wouter Hennen on 05/03/2023.
+//
+
+import SwiftUI
+import Combine
+
+class SplitViewItem: ObservableObject {
+
+    var id: AnyHashable
+    var item: NSSplitViewItem
+
+    var collapsed: Binding<Bool>
+
+    var cancellables: [AnyCancellable] = []
+
+    var observers: [NSKeyValueObservation] = []
+
+    init(child: _VariadicView.Children.Element) {
+        self.id = child.id
+        self.item = NSSplitViewItem(viewController: NSHostingController(rootView: child))
+        self.collapsed = child[SplitViewItemCollapsedViewTraitKey.self]
+        self.item.canCollapse = child[SplitViewItemCanCollapseViewTraitKey.self]
+        self.item.isCollapsed = self.collapsed.wrappedValue
+        self.observers = createObservers()
+    }
+
+    func createObservers() -> [NSKeyValueObservation] {
+        [
+            item.observe(\.isCollapsed) { item, _ in
+                self.collapsed.wrappedValue = item.isCollapsed
+            }
+        ]
+    }
+
+    func update(child: _VariadicView.Children.Element) {
+        self.item.canCollapse = child[SplitViewItemCanCollapseViewTraitKey.self]
+        DispatchQueue.main.async {
+            self.observers = []
+            self.item.animator().isCollapsed = child[SplitViewItemCollapsedViewTraitKey.self].wrappedValue
+            self.observers = self.createObservers()
+        }
+    }
+}
