+++ CodeEdit/Features/Tabs/Models/TabManager.swift
